+++ playback/src/player.rs
@@ -23,11 +23,11 @@ use crate::convert::Converter;
 use crate::core::session::Session;
 use crate::core::spotify_id::SpotifyId;
 use crate::core::util::SeqGenerator;
-use crate::decoder::{AudioDecoder, AudioError, AudioPacket, PassthroughDecoder, VorbisDecoder};
+use crate::decoder::{AudioDecoder, AudioPacket, DecoderError, PassthroughDecoder, VorbisDecoder};
 use crate::metadata::{AudioItem, FileFormat};
 use crate::mixer::AudioFilter;
 
-use crate::{NUM_CHANNELS, SAMPLES_PER_SECOND};
+use crate::{MS_PER_PAGE, NUM_CHANNELS, PAGES_PER_MS, SAMPLES_PER_SECOND};
 
 const PRELOAD_NEXT_TRACK_BEFORE_END_DURATION_MS: u32 = 30000;
 pub const DB_VOLTAGE_RATIO: f64 = 20.0;
@@ -356,7 +356,11 @@ impl Player {
     }
 
     fn command(&self, cmd: PlayerCommand) {
-        self.commands.as_ref().unwrap().send(cmd).unwrap();
+        if let Some(commands) = self.commands.as_ref() {
+            if let Err(e) = commands.send(cmd) {
+                error!("Player Commands Error: {}", e);
+            }
+        }
     }
 
     pub fn load(&mut self, track_id: SpotifyId, start_playing: bool, position_ms: u32) -> u64 {
@@ -429,7 +433,7 @@ impl Drop for Player {
         if let Some(handle) = self.thread_handle.take() {
             match handle.join() {
                 Ok(_) => (),
-                Err(_) => error!("Player thread panicked!"),
+                Err(e) => error!("Player thread Error: {:?}", e),
             }
         }
     }
@@ -505,7 +509,10 @@ impl PlayerState {
         match *self {
             Stopped | EndOfTrack { .. } | Paused { .. } | Loading { .. } => false,
             Playing { .. } => true,
-            Invalid => panic!("invalid state"),
+            Invalid => {
+                error!("PlayerState is_playing: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -530,7 +537,10 @@ impl PlayerState {
             | Playing {
                 ref mut decoder, ..
             } => Some(decoder),
-            Invalid => panic!("invalid state"),
+            Invalid => {
+                error!("PlayerState decoder: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -546,7 +556,10 @@ impl PlayerState {
                 ref mut stream_loader_controller,
                 ..
             } => Some(stream_loader_controller),
-            Invalid => panic!("invalid state"),
+            Invalid => {
+                error!("PlayerState stream_loader_controller: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -577,7 +590,10 @@ impl PlayerState {
                     },
                 };
             }
-            _ => panic!("Called playing_to_end_of_track in non-playing state."),
+            _ => {
+                error!("Called playing_to_end_of_track in non-playing state.");
+                exit(1);
+            }
         }
     }
 
@@ -610,7 +626,10 @@ impl PlayerState {
                     suggested_to_preload_next_track,
                 };
             }
-            _ => panic!("invalid state"),
+            _ => {
+                error!("PlayerState paused_to_playing: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -643,7 +662,10 @@ impl PlayerState {
                     suggested_to_preload_next_track,
                 };
             }
-            _ => panic!("invalid state"),
+            _ => {
+                error!("PlayerState playing_to_paused: invalid state");
+                exit(1);
+            }
         }
     }
 }
@@ -699,8 +721,8 @@ impl PlayerTrackLoader {
     ) -> Option<PlayerLoadedTrackData> {
         let audio = match AudioItem::get_audio_item(&self.session, spotify_id).await {
             Ok(audio) => audio,
-            Err(_) => {
-                error!("Unable to load audio item.");
+            Err(e) => {
+                error!("Unable to load audio item: {:?}", e);
                 return None;
             }
         };
@@ -768,8 +790,8 @@ impl PlayerTrackLoader {
 
             let encrypted_file = match encrypted_file.await {
                 Ok(encrypted_file) => encrypted_file,
-                Err(_) => {
-                    error!("Unable to load encrypted file.");
+                Err(e) => {
+                    error!("Unable to load encrypted file: {:?}", e);
                     return None;
                 }
             };
@@ -787,8 +809,8 @@ impl PlayerTrackLoader {
 
             let key = match self.session.audio_key().request(spotify_id, file_id).await {
                 Ok(key) => key,
-                Err(_) => {
-                    error!("Unable to load decryption key");
+                Err(e) => {
+                    error!("Unable to load decryption key: {:?}", e);
                     return None;
                 }
             };
@@ -813,12 +835,12 @@ impl PlayerTrackLoader {
             let result = if self.config.passthrough {
                 match PassthroughDecoder::new(audio_file) {
                     Ok(result) => Ok(Box::new(result) as Decoder),
-                    Err(e) => Err(AudioError::PassthroughError(e)),
+                    Err(e) => Err(DecoderError::PassthroughDecoder(e.to_string())),
                 }
             } else {
                 match VorbisDecoder::new(audio_file) {
                     Ok(result) => Ok(Box::new(result) as Decoder),
-                    Err(e) => Err(AudioError::VorbisError(e)),
+                    Err(e) => Err(DecoderError::LewtonDecoder(e.to_string())),
                 }
             };
 
@@ -830,14 +852,17 @@ impl PlayerTrackLoader {
                         e
                     );
 
-                    if self
-                        .session
-                        .cache()
-                        .expect("If the audio file is cached, a cache should exist")
-                        .remove_file(file_id)
-                        .is_err()
-                    {
-                        return None;
+                    match self.session.cache() {
+                        Some(cache) => {
+                            if cache.remove_file(file_id).is_err() {
+                                error!("Error removing file from cache");
+                                return None;
+                            }
+                        }
+                        None => {
+                            error!("If the audio file is cached, a cache should exist");
+                            return None;
+                        }
                     }
 
                     // Just try it again
@@ -849,13 +874,15 @@ impl PlayerTrackLoader {
                 }
             };
 
-            if position_ms != 0 {
-                if let Err(err) = decoder.seek(position_ms as i64) {
-                    error!("Vorbis error: {}", err);
+            let position_pcm = PlayerInternal::position_ms_to_pcm(position_ms);
+
+            if position_pcm != 0 {
+                if let Err(e) = decoder.seek(position_pcm) {
+                    error!("PlayerTrackLoader load_track: {}", e);
                 }
                 stream_loader_controller.set_stream_mode();
             }
-            let stream_position_pcm = PlayerInternal::position_ms_to_pcm(position_ms);
+            let stream_position_pcm = position_pcm;
             info!("<{}> ({} ms) loaded", audio.name, audio.duration);
 
             return Some(PlayerLoadedTrackData {
@@ -912,7 +939,8 @@ impl Future for PlayerInternal {
                             start_playback,
                         );
                         if let PlayerState::Loading { .. } = self.state {
-                            panic!("The state wasn't changed by start_playback()");
+                            error!("The state wasn't changed by start_playback()");
+                            exit(1);
                         }
                     }
                     Poll::Ready(Err(_)) => {
@@ -976,47 +1004,67 @@ impl Future for PlayerInternal {
                     ..
                 } = self.state
                 {
-                    let packet = decoder.next_packet().expect("Vorbis error");
+                    match decoder.next_packet() {
+                        Ok(packet) => {
+                            if !passthrough {
+                                if let Some(ref packet) = packet {
+                                    match packet.samples() {
+                                        Ok(samples) => {
+                                            *stream_position_pcm +=
+                                                (samples.len() / NUM_CHANNELS as usize) as u64;
+                                            let stream_position_millis =
+                                                Self::position_pcm_to_ms(*stream_position_pcm);
 
-                    if !passthrough {
-                        if let Some(ref packet) = packet {
-                            *stream_position_pcm +=
-                                (packet.samples().len() / NUM_CHANNELS as usize) as u64;
-                            let stream_position_millis =
-                                Self::position_pcm_to_ms(*stream_position_pcm);
-
-                            let notify_about_position = match *reported_nominal_start_time {
-                                None => true,
-                                Some(reported_nominal_start_time) => {
-                                    // only notify if we're behind. If we're ahead it's probably due to a buffer of the backend and we're actually in time.
-                                    let lag = (Instant::now() - reported_nominal_start_time)
-                                        .as_millis()
-                                        as i64
-                                        - stream_position_millis as i64;
-                                    lag > Duration::from_secs(1).as_millis() as i64
+                                            let notify_about_position =
+                                                match *reported_nominal_start_time {
+                                                    None => true,
+                                                    Some(reported_nominal_start_time) => {
+                                                        // only notify if we're behind. If we're ahead it's probably due to a buffer of the backend and we're actually in time.
+                                                        let lag = (Instant::now()
+                                                            - reported_nominal_start_time)
+                                                            .as_millis()
+                                                            as i64
+                                                            - stream_position_millis as i64;
+                                                        lag > Duration::from_secs(1).as_millis()
+                                                            as i64
+                                                    }
+                                                };
+                                            if notify_about_position {
+                                                *reported_nominal_start_time = Some(
+                                                    Instant::now()
+                                                        - Duration::from_millis(
+                                                            stream_position_millis as u64,
+                                                        ),
+                                                );
+                                                self.send_event(PlayerEvent::Playing {
+                                                    track_id,
+                                                    play_request_id,
+                                                    position_ms: stream_position_millis as u32,
+                                                    duration_ms,
+                                                });
+                                            }
+                                        }
+                                        Err(e) => {
+                                            error!("PlayerInternal poll: {}", e);
+                                            exit(1);
+                                        }
+                                    }
                                 }
-                            };
-                            if notify_about_position {
-                                *reported_nominal_start_time = Some(
-                                    Instant::now()
-                                        - Duration::from_millis(stream_position_millis as u64),
-                                );
-                                self.send_event(PlayerEvent::Playing {
-                                    track_id,
-                                    play_request_id,
-                                    position_ms: stream_position_millis as u32,
-                                    duration_ms,
-                                });
+                            } else {
+                                // position, even if irrelevant, must be set so that seek() is called
+                                *stream_position_pcm = duration_ms.into();
                             }
-                        }
-                    } else {
-                        // position, even if irrelevant, must be set so that seek() is called
-                        *stream_position_pcm = duration_ms.into();
-                    }
 
-                    self.handle_packet(packet, normalisation_factor);
+                            self.handle_packet(packet, normalisation_factor);
+                        }
+                        Err(e) => {
+                            error!("PlayerInternal poll: {}", e);
+                            exit(1);
+                        }
+                    }
                 } else {
-                    unreachable!();
+                    error!("PlayerInternal poll: Invalid PlayerState");
+                    exit(1);
                 };
             }
 
@@ -1065,11 +1113,11 @@ impl Future for PlayerInternal {
 
 impl PlayerInternal {
     fn position_pcm_to_ms(position_pcm: u64) -> u32 {
-        (position_pcm * 10 / 441) as u32
+        (position_pcm as f64 * MS_PER_PAGE) as u32
     }
 
     fn position_ms_to_pcm(position_ms: u32) -> u64 {
-        position_ms as u64 * 441 / 10
+        (position_ms as f64 * PAGES_PER_MS) as u64
     }
 
     fn ensure_sink_running(&mut self) {
@@ -1080,8 +1128,8 @@ impl PlayerInternal {
             }
             match self.sink.start() {
                 Ok(()) => self.sink_status = SinkStatus::Running,
-                Err(err) => {
-                    error!("Fatal error, could not start audio sink: {}", err);
+                Err(e) => {
+                    error!("{}", e);
                     exit(1);
                 }
             }
@@ -1103,8 +1151,8 @@ impl PlayerInternal {
                             callback(self.sink_status);
                         }
                     }
-                    Err(err) => {
-                        error!("Fatal error, could not stop audio sink: {}", err);
+                    Err(e) => {
+                        error!("{}", e);
                         exit(1);
                     }
                 }
@@ -1151,7 +1199,10 @@ impl PlayerInternal {
                 self.state = PlayerState::Stopped;
             }
             PlayerState::Stopped => (),
-            PlayerState::Invalid => panic!("invalid state"),
+            PlayerState::Invalid => {
+                error!("PlayerInternal handle_player_stop: invalid state");
+                exit(1);
+            }
         }
     }
 
@@ -1317,8 +1368,8 @@ impl PlayerInternal {
                         }
                     }
 
-                    if let Err(err) = self.sink.write(&packet, &mut self.converter) {
-                        error!("Fatal error, could not write audio to audio sink: {}", err);
+                    if let Err(e) = self.sink.write(&packet, &mut self.converter) {
+                        error!("{}", e);
                         exit(1);
                     }
                 }
@@ -1337,7 +1388,8 @@ impl PlayerInternal {
                         play_request_id,
                     })
                 } else {
-                    unreachable!();
+                    error!("PlayerInternal handle_packet: Invalid PlayerState");
+                    exit(1);
                 }
             }
         }
@@ -1458,7 +1510,10 @@ impl PlayerInternal {
                 play_request_id,
                 position_ms,
             }),
-            PlayerState::Invalid { .. } => panic!("Player is in an invalid state."),
+            PlayerState::Invalid { .. } => {
+                error!("PlayerInternal handle_command_load: invalid state");
+                exit(1);
+            }
         }
 
         // Now we check at different positions whether we already have a pre-loaded version
@@ -1474,24 +1529,30 @@ impl PlayerInternal {
             if previous_track_id == track_id {
                 let mut loaded_track = match mem::replace(&mut self.state, PlayerState::Invalid) {
                     PlayerState::EndOfTrack { loaded_track, .. } => loaded_track,
-                    _ => unreachable!(),
+                    _ => {
+                        error!("PlayerInternal handle_command_load: Invalid PlayerState");
+                        exit(1);
+                    }
                 };
 
-                if Self::position_ms_to_pcm(position_ms) != loaded_track.stream_position_pcm {
+                let position_pcm = Self::position_ms_to_pcm(position_ms);
+
+                if position_pcm != loaded_track.stream_position_pcm {
                     loaded_track
                         .stream_loader_controller
                         .set_random_access_mode();
-                    let _ = loaded_track.decoder.seek(position_ms as i64); // This may be blocking.
-                                                                           // But most likely the track is fully
-                                                                           // loaded already because we played
-                                                                           // to the end of it.
+                    if let Err(e) = loaded_track.decoder.seek(position_pcm) {
+                        // This may be blocking.
+                        error!("PlayerInternal handle_command_load: {}", e);
+                    }
                     loaded_track.stream_loader_controller.set_stream_mode();
-                    loaded_track.stream_position_pcm = Self::position_ms_to_pcm(position_ms);
+                    loaded_track.stream_position_pcm = position_pcm;
                 }
                 self.preload = PlayerPreload::None;
                 self.start_playback(track_id, play_request_id, loaded_track, play);
                 if let PlayerState::Invalid = self.state {
-                    panic!("start_playback() hasn't set a valid player state.");
+                    error!("start_playback() hasn't set a valid player state.");
+                    exit(1);
                 }
                 return;
             }
@@ -1515,11 +1576,16 @@ impl PlayerInternal {
         {
             if current_track_id == track_id {
                 // we can use the current decoder. Ensure it's at the correct position.
-                if Self::position_ms_to_pcm(position_ms) != *stream_position_pcm {
+                let position_pcm = Self::position_ms_to_pcm(position_ms);
+
+                if position_pcm != *stream_position_pcm {
                     stream_loader_controller.set_random_access_mode();
-                    let _ = decoder.seek(position_ms as i64); // This may be blocking.
+                    if let Err(e) = decoder.seek(position_pcm) {
+                        // This may be blocking.
+                        error!("PlayerInternal handle_command_load: {}", e);
+                    }
                     stream_loader_controller.set_stream_mode();
-                    *stream_position_pcm = Self::position_ms_to_pcm(position_ms);
+                    *stream_position_pcm = position_pcm;
                 }
 
                 // Move the info from the current state into a PlayerLoadedTrackData so we can use
@@ -1558,12 +1624,14 @@ impl PlayerInternal {
                     self.start_playback(track_id, play_request_id, loaded_track, play);
 
                     if let PlayerState::Invalid = self.state {
-                        panic!("start_playback() hasn't set a valid player state.");
+                        error!("start_playback() hasn't set a valid player state.");
+                        exit(1);
                     }
 
                     return;
                 } else {
-                    unreachable!();
+                    error!("PlayerInternal handle_command_load: Invalid PlayerState");
+                    exit(1);
                 }
             }
         }
@@ -1581,17 +1649,23 @@ impl PlayerInternal {
                     mut loaded_track,
                 } = preload
                 {
-                    if Self::position_ms_to_pcm(position_ms) != loaded_track.stream_position_pcm {
+                    let position_pcm = Self::position_ms_to_pcm(position_ms);
+
+                    if position_pcm != loaded_track.stream_position_pcm {
                         loaded_track
                             .stream_loader_controller
                             .set_random_access_mode();
-                        let _ = loaded_track.decoder.seek(position_ms as i64); // This may be blocking
+                        if let Err(e) = loaded_track.decoder.seek(position_pcm) {
+                            // This may be blocking
+                            error!("PlayerInternal handle_command_load: {}", e);
+                        }
                         loaded_track.stream_loader_controller.set_stream_mode();
                     }
                     self.start_playback(track_id, play_request_id, *loaded_track, play);
                     return;
                 } else {
-                    unreachable!();
+                    error!("PlayerInternal handle_command_load: Invalid PlayerState");
+                    exit(1);
                 }
             }
         }
@@ -1697,7 +1771,9 @@ impl PlayerInternal {
             stream_loader_controller.set_random_access_mode();
         }
         if let Some(decoder) = self.state.decoder() {
-            match decoder.seek(position_ms as i64) {
+            let position_pcm = Self::position_ms_to_pcm(position_ms);
+
+            match decoder.seek(position_pcm) {
                 Ok(_) => {
                     if let PlayerState::Playing {
                         ref mut stream_position_pcm,
@@ -1708,10 +1784,10 @@ impl PlayerInternal {
                         ..
                     } = self.state
                     {
-                        *stream_position_pcm = Self::position_ms_to_pcm(position_ms);
+                        *stream_position_pcm = position_pcm;
                     }
                 }
-                Err(err) => error!("Vorbis error: {:?}", err),
+                Err(e) => error!("PlayerInternal handle_command_seek: {}", e),
             }
         } else {
             warn!("Player::seek called from invalid state");
@@ -1954,7 +2030,9 @@ struct Subfile<T: Read + Seek> {
 
 impl<T: Read + Seek> Subfile<T> {
     pub fn new(mut stream: T, offset: u64) -> Subfile<T> {
-        stream.seek(SeekFrom::Start(offset)).unwrap();
+        if let Err(e) = stream.seek(SeekFrom::Start(offset)) {
+            error!("Subfile new Error: {}", e);
+        }
         Subfile { stream, offset }
     }
 }